Make shebang for python an absolute command path
These scripts use 'env' as an interpreter. For the rpm runtime dependency
detection to work, the shebang #!/usr/bin/env python  needs to be patched into
 #!/usr/bin/python  otherwise the package dependency generator merely adds a
dependency on /usr/bin/env rather than the actual interpreter /usr/bin/python.
Alternatively, if the file should not be executed, then ensure that it is not
marked as executable or don't install it in a path that is reserved for
executables.